DF_Web/requirements.txt
# Core API Framework
fastapi
uvicorn[standard]

# AI/ML Dependencies
torch
torchvision
pillow
numpy
opencv-python

# HTTP Requests (for model download)
requests

# Production server
gunicorn

# Python multipart (for file uploads)
python-multipart